+++ bc-react/package.json
@@ -6,7 +6,6 @@
     "create-react-class": "^15.5.4",
     "flux": "^3.1.2",
     "jquery": "^3.2.1",
-    "moment": "^2.18.1",
     "moment-timezone": "^0.5.13",
     "node-sass-chokidar": "^0.0.2",
     "npm-run-all": "^4.0.2",
