+++ bc-react/src/components/Admin/AdminTable.js
@@ -204,12 +204,21 @@ class AdminTable extends Component {
           <td className="icon_td">{this.editIcon()}</td>
         </tr>
     )}else if(this.props.eventTable){
+
+        let weekday = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"]
+        let months = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December']
+        let temp = this.state.event.date.split('T')
+        let date = temp[0].split('-')
+        let month = months[new Date(date).getMonth()]
+        let day = new Date(date).getDate()
+        let year = new Date(temp).getFullYear()
+
       return(
       <tr className={this.state.className}>
         <td>
           <input
             name='date'
-            value={this.state.event.date}
+            value={`${month} ${day}, ${year}`}
             onChange={this.handleEdit.bind(this)}
             disabled={this.state.readOnly} /></td>
         <td>
